LayoutTests/fast/webgpu/bind-group-duplicate-binding-expected.txt

CONSOLE MESSAGE: GPUDevice.createBindGroup: Binding 0 is duplicated in the bind group descriptor
